session2/fis/requirements.txt
gradio>=4.0.0
pandas>=2.0.0
pyyaml>=6.0.0
wordcloud>=1.9.0
matplotlib>=3.7.0
numpy>=1.24.0 
kaleido>=0.2.1
plotly>=6.0.1
llama-index-embeddings-huggingface